+++ DeepMuon/dataset/HailingData.py
@@ -15,7 +15,7 @@ import numba
 
 import torch
 from torch.utils.data import Dataset
-torch.set_default_tensor_type(torch.DoubleTensor)
+torch.set_default_tensor_type(torch.FloatTensor)
 
 
 @numba.jit
@@ -126,11 +126,11 @@ class HailingDataset_DirectV3(Dataset):
             oper = np.unique(np.random.randint(0, 4, np.random.randint(0, 4)))
             for oper_i in range(len(oper)):
                 image, label = self.augmentation[oper[oper_i]](image, label)
-        image = torch.from_numpy(image.copy())
+        image = torch.from_numpy(image.copy()).type(torch.FloatTensor)
         # image = torch.permute(image, (3, 0, 1, 2))
         image = torch.permute(image, (3, 2, 0, 1))
         image[1:, :, :, :] = 0.0001*image[1:, :, :, :]
-        label = torch.from_numpy(label)
+        label = torch.from_numpy(label).type(torch.FloatTensor)
         return image, label
 
     def __Init(self):
@@ -175,11 +175,11 @@ class HailingDataset_Direct2(Dataset):
             oper = np.unique(np.random.randint(0, 4, np.random.randint(0, 4)))
             for oper_i in range(len(oper)):
                 image, label = self.augmentation[oper[oper_i]](image, label)
-        image = torch.from_numpy(image.copy())
+        image = torch.from_numpy(image.copy()).type(torch.FloatTensor)
         image = torch.permute(image, (3, 0, 1, 2))
         # image = torch.permute(image, (3, 2, 0, 1))
         image[1:, :, :, :] = 0.0001*image[1:, :, :, :]
-        label = torch.from_numpy(label)
+        label = torch.from_numpy(label).type(torch.FloatTensor)
         return image, label
 
     def __Init(self):
