+++ DeepMuon/tools/AirFunc.py
@@ -16,7 +16,7 @@ import matplotlib.pyplot as plt
 
 import torch
 from torch import nn
-torch.set_default_tensor_type(torch.DoubleTensor)
+torch.set_default_tensor_type(torch.FloatTensor)
 
 
 def exclude_key(dictionary: dict, del_key: str = 'type'):
